Finished github login
