+++ services/api/auth.js
@@ -16,6 +16,9 @@ const login = async (type, options) => {
     case 'google':
       idToken = await handleGoogle(options)
       break
+    case 'github':
+      idToken = await handleGithub(options)
+      break
     default:
       throw new Error('Unknown login type expected one of ["email", "google"]')
   }
@@ -43,6 +46,19 @@ const handleEmailAndPassword = async ({ email, password }) => {
 const handleGoogle = async ({ googleIdToken }) => {
   // Create firebase credentials with google id token
   const credential = firebase.auth.GoogleAuthProvider.credential(googleIdToken)
+  // Try to login to user
+  const { user } = await firebase.auth().signInWithCredential(credential)
+  // Get token for user
+  return user.getIdToken(true)
+}
+
+const handleGithub = async ({ clientId, code }) => {
+  // Get auth token for github with temp code given by Github callback
+  const githubAuthUrl = 'http://localhost:3000/api/auth/github'
+  const options = { method: 'POST', body: { clientId, code } }
+  const { access_token: githubIdToken } = await new Fetch(githubAuthUrl, options).json()
+  // Create firebase credentials with github id token
+  const credential = firebase.auth.GithubAuthProvider.credential(githubIdToken)
   // Try to login to user
   const { user } = await firebase.auth().signInWithCredential(credential)
   // Get token for user
